requirements.txt
torch>=2.0.0
numpy>=1.24.0
python-chess>=1.999
matplotlib>=3.7.0
psutil>=5.9.0
tqdm>=4.65.0